ABCD functional, still need to fix memory leaks and tweaks

+++ src/AlphaBetaAction.cpp
@@ -1,7 +1,7 @@
 #include "AlphaBetaAction.h"
 #include "AlphaBetaUnit.h"
 
-AlphaBetaAction::AlphaBetaAction(AlphaBetaUnit * punit, AlphaBetaUnit * ptarget, sc2::Point2D pposition, float pdistance, AlphaBetaActionType ptype, long ptime) {
+AlphaBetaAction::AlphaBetaAction(AlphaBetaUnit * punit, AlphaBetaUnit * ptarget, sc2::Point2D pposition, float pdistance, AlphaBetaActionType ptype, float ptime) {
     unit = punit;
     target = ptarget;
     position = pposition;
